refactor: 관리자 페이지 로그아웃 버튼 수정
